python/rag/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2

langchain==0.3.25
langchain-openai==0.3.17
langchain-pinecone==0.2.6
openai==1.77.0

pinecone==6.0.2

mysql-connector-python==9.3.0

beautifulsoup4==4.13.4
selenium==4.31.0

pillow==11.2.1
opencv-python==4.11.0.86

boto3==1.38.11

python-dotenv==1.1.0
requests==2.32.3